+++ lib/montycat.dart
@@ -18,4 +18,5 @@ export 'src/semantic.dart'
 export 'src/classes/inmemory.dart' show KeyspaceInMemory;
 export 'src/classes/persistent.dart' show KeyspacePersistent;
 export 'src/utils.dart' show SubscriptionHandle;
+export 'src/tls.dart' show TlsSettings, TlsVerificationException;
 export 'src/pool.dart' show PoolConfig, closeAllPools;